refactor: change some IDs and DTO fields from Integer to Long (still errors in some services)

+++ Backend/src/main/java/com/sena/urbantracker/routes/repository/IRoute.java
@@ -3,5 +3,5 @@ package com.sena.urbantracker.routes.repository;
 import com.sena.urbantracker.routes.model.entity.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRoute extends JpaRepository<Route, Integer> {
+public interface IRoute extends JpaRepository<Route, Long> {
 }
